+++ CHANGELOG.md
@@ -6,11 +6,12 @@
 - Street Detector
 - Port Scanner
 - Web Console
-- Add support for variable display timeout on TWatch S3
+- Improve WiFi Connection logic
 
 
 ## 1.4.5
-
+- Forward printf logs from managers to terminal view
+- Added support for variable display timeout on TWatch S3
 - Added "gpsinfo" command to get GPS information
 - make sure .csv files are closed with generic stop command
 - Removed wifi warnings and color codes from logs to clean up the output